yogita_bot/install/requirements.txt
pyTelegramBotAPI==4.15.2
python-dotenv==1.0.0
peewee==3.17.0
pandas==2.3.3
openpyxl==3.1.5
requests==2.32.5
PyMySQL==1.1.0
mysql-connector-python==8.1.0